#engine : refactor domains
