+++ test/domains/field/field.test.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 chai.should();
 
-const Field = require('../../src/domain/field');
+const Field = require('../../../src/domains/field/field');
 
 describe('Field', () => {
   it('should have no mowers when field created as empty', () => {
